REFACTOR - Refactoring View.js files into their own folder so page subcomponents are organized in their own respective folder
